+++ lib/td_df_web/hypermedia/hypermedia_collection.ex
@@ -1,5 +1,0 @@
-defmodule TdDfWeb.Hypermedia.HypermediaCollection do
-  @moduledoc false
-
-  defstruct [:collection_hypermedia, :collection]
-end