+++ lib/td_df_web/hypermedia/link.ex
@@ -1,5 +1,0 @@
-defmodule TdDfWeb.Hypermedia.Link do
-  @moduledoc false
-
-  defstruct [:action, :path, :method, :schema]
-end